First shot at specifying an interchange format for ABI change reports
Okay, this is far from being finished and is work in progress.  I am
basically going through the code that emits the text report for the
different kinds of diff nodes and I am generalizing that into an
extensible format.  The schema of the format is expressed in
RELAX-NG. For now the resulting format closely matches what is emitted
in current text reporter.

	* doc/manuals/ifac-schema.rng: New file containing the relax-ng
	schema of the xml report.
	* doc/manuals/libabigail-xml-report.rst: New doc that talks about
	the schema of the XML format.  For now it just includes the
	schema.  But it's intended to explain the format in plain english
	too.
	* doc/manuals/libabigail-concepts.rst: Add a paragraph about
	report formats and link the schema for the XML format in there.
	* doc/manuals/Makefile.am: Add the new files to the source
	distribution.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>