[abipkgdiff] End all branches of get_soname() finish with a return statement
The compiler complains that there are some branches in the
get_soname() function that don't end up with a return statement.  For
instance, there is no return statement right before the end of the
function, in case the outermost for loop is not taken.

This patch fixes that by having just one return statement at the end
of the function.

	* tools/abipkgdiff.cc (get_soname): Have just one return statement
	at the end of this function.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>