corpus: is_empty: consider actual translation unit contents
A corpus with completely filtered out symbols (exhaustive whitelist),
still contains compilation units, but they are empty. A list of empty
translation units shall be considered empty as no entries need to be
considered. That is useful to skip empty corpora when writing out the
xml for them.

Hence, teach is_empty() to have a look at the actual translation units.

	* src/abg-corpus.cc (corpus::is_empty): consider a list of
	  empty members to be empty.

Signed-off-by: Matthias Maennich <maennich@google.com>