abg-dwarf-reader.cc: Sort translation units before processing them.
A common source of kernel ABI churn is the movement of type declarations
between abi-instr XML elements.

While it may be possible to fix this in the longer term by moving
types to a dedicated section of ABI XML, a short-term fix would make
reviewing ABI changes less painful.

It is suspected that translation unit order within a corpus may
contribute to this issue. This commit imposes a deterministic order.

Some of the variations in test output may point to latent bugs.

Signed-off-by: Giuliano Procida <gprocida@google.com>