Tidy new line logic in represent function.
The represent function has a lot of cases with similar logic. In each
case there is a three-way choice:

    1 diff text already emitted, continue on a new line
    2 diff text already emitted, continue on the same line
    3 diff text not yet emitted

This patch reorders the choices into the above order (instead of 1, 3,
2) and moves the update of the emitted state variable to be adjacent
to the conditionals (instead of at the end of the block).

The result is a little easier to read (unvarying code preceeds varying
code in each case's block). This patch also simplifies the catch-all
logic and improves the comments there.

There are no changes to behaviour.

	* src/abg-reporter-priv.cc (represent): In the var_diff_sptr
	overload, reorder the 3-way choices (as begin_with_and, emitted,
	!emitted) in all the blocks that do this, move emitted = true
	up in the same blocks, clean up catch-all conditional logic
	and comments at the end of the function.

Signed-off-by: Giuliano Procida <gprocida@google.com>