Factor out suppression property string parsing.
This patch introduces a helper function to get a string value from an
ini config property, with error checking. All string look-ups now use
this.

Note that the function is named "read" and it is the first of many
overloads of the same name, to allow type-driven parsing (analogous to
operator>>).

There are no behavioural changes, except as follows:

- if the string value is present but empty, it will be used, rather
  than ignored; in all current cases, this makes no difference
- if both "drop_artifact" and "drop" properties are (spuriously)
  present and "drop_artifact" is (erroneously) not a simple property,
  then the "drop" property will be ignored whereas previously it would
  have been examined

The commit also adds a placeholder for error reporting in the case
that the property found is not a simple string property.

	* src/abg-suppression.cc (read) Add new overloaded
	function. This overload reads a string value from a property.
	(read_type_suppression): Refactor code to fetch strings using
	read helper function. Move fetches to point of use in code so
	that setting logic is conditional on fetch and parse success.
	Make sure drop artifact logic happens late in function as it
	will depend on other fields in suppression being populated.
	(read_function_suppression): Ditto.
	(read_variable_suppression): Ditto.
	(read_file_suppression): Ditto.

Signed-off-by: Giuliano Procida <gprocida@google.com>