Simplify reach kind filtering, with error handling.
Reach kinds are enumerated and there is no obvious "not present" value
that can be used by default. The current code uses an extra boolean
within type_suppression to determine whether or not to actually
consider the reach kind therein.

Misspelt reach kinds are silently recorded as "direct" which is
actually does mean "no filtering" as currently implemented.

This change eliminates the extra boolean (and associated getter and
setter), instead using a dedicated value "unspecified" in the enum.

In the case of an unrecognised reach kind being provided an error will
be returned and the property will continue be ignored as if absent.

	* include/abg-suppression.h (type_suppression::reach_kind):
	Add UNSPECIFIED_REACH_KIND enumerator. (type_suppression):
	Drop consider_reach_kind getter and setter methods.
	* src/abg-suppression-priv.h (type_suppression::priv): Drop
	consider_reach_kind_ member. Default reach_kind_ to
	UNSPECIFIED_REACH_KIND instead of DIRECT_REACH_KIND.
	* src/abg-suppression.cc (string_to_reach_kind): New helper
	function adapted from old read_suppression_reach_kind, with error
	checking. (read) New overload to parse properties to
	type_suppression::reach_kind values.
	(type_suppression): Drop consider_reach_kind getter and setter
	methods. (type_suppression::set_reach_kind): Update comments.
	(type_suppression::get_reach_kind): Update comments.
	(type_suppression::suppresses_diff): Check for
	UNSPECIFIED_REACH_KIND instead of !consider_reach_kind.
	(read_type_suppression): Use read overload instead of old
	read_suppression_reach_kind to parse "reach_kind" properties.

Signed-off-by: Giuliano Procida <gprocida@google.com>