Simplify type kind filtering, with error handling.
Type kinds are enumerated and there is no "not present" value that can
be used by default. The current code instead uses an extra boolean
within type_suppression to determine whether or not to actually
consider the type kind therein.

Misspelt type kinds are currently recorded as "unknown" but this is
handled in the same way as "class".

This change eliminates the extra boolean (and associated getter and
setter), instead using a dedicated value "unspecified" in the enum.

In the case of an unrecognised type kind being provided an error will
be returned and the property will now be ignored (instead of being
treated as "class").

	* include/abg-suppression.h (type_suppression::type_kind):
	Rename UNKNOWN_TYPE_KIND enumerator to UNSPECIFIED_TYPE_KIND.
	(type_suppression): Drop consider_type_kind getter and setter
	methods.
	* src/abg-suppression-priv.h (type_suppression::priv): Drop
	consider_type_kind_ member. Default type_kind_ to
	UNSPECIFIED_TYPE_KIND instead of CLASS_TYPE_KIND.
	* src/abg-suppression.cc (string_to_type_kind): New helper
	function adapted from old read_type_kind_string, with error
	checking. (read) New overload to parse properties to
	type_suppression::type_kind values.
	(type_suppression): Drop consider_type_kind getter and setter
	methods. (type_suppression::set_type_kind): Update comments.
	(type_suppression::get_type_kind): Update comments.
	(suppression_matches_type_no_name): Remove UNKNOWN_TYPE_KIND
	case and add case for UNSPECIFIED_TYPE_KIND equivalent to old
	!consider_type_kind.  (read_type_suppression): Use read
	overload instead of old read_type_kind_string to parse
	"type_kind" properties.

Signed-off-by: Giuliano Procida <gprocida@google.com>