Factor out boolean parsing, with error checking.
This patch factors out boolean parsing into a helper function. The
values "no" and "false" are now explicitly recognised. Other
unrecognised values are treated as parse failures and there is now a
placholder where an error message could be emitted.

As all current callers are oblivious to the distinction between parse
failure and false, there should be no behavioural changes.

	src/abg-suppression.cc (string_to_boolean): Convert a string
	to a boolean, returning an error if needed.
	(read): New bool overload, to get a boolean from a property.
	(read_type_suppression): Use read for drop_artifact and drop
	properties.
	(read_function_suppression): Use read for
	allow_other_aliases, drop_artifact and drop properties.
	(read_variable_suppression): Use read for drop_artifact,
	drop and allow_other_aliases properties.
	(read_file_suppression): Use read for drop_artifact
	and drop properties.

Signed-off-by: Giuliano Procida <gprocida@google.com>