Rename insertion_range types.
This is a prelude for a change to make the type representing data
member offsets stand independently of the type representing offset
ranges.

The type insertion_range is currently named for its purpose, rather
than its essence which is an abstract representation of a range of
offsets. Similarly, boundary is the purpose of an offset in the
context of a range.

This patch renames types and typedefs as follows and updates comments.

- insertion_range -> offset_range
- boundary -> offset
- integer_boundary -> integer_offset
- fn_call_expr_boundary -> fn_call_expr_offset
- insertion_ranges -> offset_ranges
- insertion_range_sptr -> offset_range_sptr
- boundary_sptr -> offset_sptr
- integer_boundary_sptr -> integer_offset_sptr
- fn_call_expr_boundary_sptr -> fn_call_expr_offset_sptr

There are no behavioural changes.

	* include/abg-suppression.h: Rename insertion_range class,
	member classes and associated typedefs. Update comments.
	* src/abg-suppression-priv.h: Ditto.
	* src/abg-suppression.cc: Ditto.

Signed-off-by: Giuliano Procida <gprocida@google.com>