Factor out offset range parsing, with error checking.
This changes factors out and adds error checking (and placeholders for
error reporting) to the parsing code for offset range (offset pair)
values, as found in has_data_member_inserted_between and
has_data_members_inserted_between properties.

Behavioural changes are as follows:

- if a has_data_member_inserted_between property exists but is of the
  wrong kind, this is now treated as an error instead of being skipped
  silently

Note that there was an ABG_ASSERT, removed as part of this commit, in
the code for has_data_members_inserted_between. It was retesting
an expression that was part of the surrounding conditional.

	src/abg-suppression.cc (property_value_to_offset_range): Add
	helper to read an offset range out of property value,
	returning an error if needed.
	(read): New offset_range_sptr overload, to read an offset
	range from a property, uses property_value_to_offset_range.
	(read_type_suppression): Use new read overload for
	has_data_member_inserted_between properties. Use
	property_value_to_offset_range helper for
	has_data_members_inserted_between properties.

Signed-off-by: Giuliano Procida <gprocida@google.com>