Remove is_well_formed state variable.
Parsing has_data_members_inserted_between properties requiries
decomposing a non-trivial property structure. If a subvalue of the
property is not as expected, the code sets the is_well_formed state
variable to false and breaks out of a processing loop.

However, due to a following conditional, this is equivalent to just
returning false from read_type_suppression and skipping the entire
suppression specification.

This commit removes the variable and simplifies the code.

Note that this early exit behaviour is peculiar to the
has_data_member* properties and will likely change.

	* src/abg-suppression.cc (read_type_suppression): Drop the
	is_well_formed variable and its associated logic. Just return
	false if there is the has_data_members_inserted_between
	property has structural issues.

Signed-off-by: Giuliano Procida <gprocida@google.com>