More input validation in string_to_parameter_spec.
Suppression function parameter specifications look like:

optional whitespace
optional "'" + positive integer
optional whitespace
optional "/" + regex + "/" or string

where at least one of the integer, regex or string are provided.

The code works by scanning through the input but is missing a few
checks to avoid going off the end. It also doesn't complain if the
second "/" is missing, is followed by junk or if the regex doesn't
compile.

This patch resolves these issues and reports an error in all cases
that a parameter specification could not be constructed.

Placeholders for error reporting have been added.

	* src/abg-suppression.cc (string_to_parameter_spec): Don't go
	OOB when checking for "'" or "/" characters. Track when we are
	between "/" delimiters and use this to report a missing end
	delimiter. Report trailing junk. Allow empty (delimited)
	regexes. Report failure if regex doesn't compile.

Signed-off-by: Giuliano Procida <gprocida@google.com>