Factor out parameter spec parsing, remove assert.
The "parameter" property is currently the only one that can be
meaningfully repeated in a suppression specification. Each parameter
should be a simple string and the code will currently abort if it
finds a composite property.

This commit removes the assertion and reuses the existing read
overload that parses a property into a string. Bad parameter spec
properties will now be reported (and ignored) in all cases.

	* src/abg-suppression (read_function_suppression): Replace
	"parameter" parsing code with standard call to overloaded read
	function.

Signed-off-by: Giuliano Procida <gprocida@google.com>