Factor out list parsing, with error handling.
This commit adds a read list overload that abstracts obtaining a list
of strings from either a simple or a list property. It returns an
error if the property is not of one of these kinds.

The read list overload is used to implement a read vector overload
which in turn is used to simplify the parsing of changed enumerators.

The commit also separates the check to see if the type kind is set to
"enum" from the adjacent parsing logic and adds a placeholder for a
warning message.

There are no behavioural changes.

	* src/abg-suppression.cc (read): New overload to read lists of
	strings from a property, writing to an iterator to support
	multiple container types.
	(read): New overload to get a vector of strings from a
	property.
	(read_type_suppression): Read changed_enumerators property
	using read_vector. Separate ENUM_TYPE_KIND check from parsing.

Signed-off-by: Giuliano Procida <gprocida@google.com>