Factor out data member insertion mutators.
The has_data_member* properties all result in offset ranges being
added to a suppression specification via a temporary vector of ranges
which is then assigned to the specification.

This patch add ranges to the specification as they are parsed and
further simplifies read_type_suppression by moving the remaining range
logic to new type_suppression mutator methods.

There are no behavioural changes.

	* include/abg-suppression.h
	(type_suppression): Declare new methods
	add_data_member_insertion_offset,
	add_data_member_insertion_range and
	add_data_member_insertion_ranges.
	* src/abg-suppression.cc: Define new type_suppression
	add_data_member_insertion_offset,
	add_data_member_insertion_range and
	add_data_member_insertion_ranges methods with code mostly
	taken from read_type_suppression.
	(read_type_suppression): Drop insert_ranges variable and
	replace code manipulating it with calls to new mutator
	methods.

Signed-off-by: Giuliano Procida <gprocida@google.com>