Pass bind_function_type_life_time a complete type.
Calling bind_function_type_life_time with an incomplete function type
at one point contributed to incomplete pretty representations of types
being generated and cached and then later used in reports. This was
fixed by making sure the caching code was aware of this.

This patch moves the call to this function to after the type is
complete. It may permit simplification of the caching logic.

	* src/abg-reader.cc: Move call to bind_function_type_life_time
	to end of function.

Signed-off-by: Giuliano Procida <gprocida@google.com>