Pass argument to diff_to_be_reported by reference.
diff_to_be_reported is a helper method within the reporter classes.
Its argument is a raw pointer, which is unusual in the code base.
Many callers have a reference and take its address before passing it.
Others have some flavour of shared_ptr and obtain the held pointer
with get().

A null diff pointer is currently treated the same as a diff without
differences but should be a program error.

This patch changes the code to pass references. The address-of
operators are removed and get() is replaced with *.

Unexpected null pointers will cause program aborts.

Signed-off-by: Giuliano Procida <gprocida@google.com>