Support constructing opaque types for enums
Now that we are to support opaque types for enums, we ought to support
building opaque types for enums as well, to mimic the opaque type
design pattern used in C, when private types are specified with type
suppression specifications.

The core of this change is to make get_opaque_version_of_type work for
enums as well, just like what it does for classes.  Note that that
function doesn't support opaque unions yet either.  That should be
added at a later point in time, I guess.

	* src/abg-dwarf-reader.cc
	(build_internal_underlying_enum_type_name)
	(build_enum_underlying_type): Factorize these functions out of ...
	(build_enum_type): ... here.
	(get_opaque_version_of_type): Make this handle enums as well.  So
	make its return type be type_or_decl_base_sptr, rather than just
	class_or_union_sptr as it used to be.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>
Signed-off-by: Giuliano Procida <gprocida@google.com>